+++ packages/fighting-design/_hooks/use-message/index.ts
@@ -1,11 +1,9 @@
-import messageVue from '../../message/src/message.vue'
-import notificationVue from '../../notification/src/notification.vue'
 import { render, createVNode } from 'vue'
 import { useMassageManage, useRun } from '..'
 import { isString } from '../../_utils'
 import { FIGHTING_TYPE } from '../../_tokens'
 import type { FightingType } from '../../_interface'
-import type { ComponentInternalInstance, VNode } from 'vue'
+import type { ComponentInternalInstance, VNode, Component } from 'vue'
 import type {
   MessageInstance,
   MessageFn,
@@ -19,50 +17,58 @@ export * from './interface.d'
 
 export const massageManage = useMassageManage()
 
-export const useMessage = (target: 'message' | 'notification'): UseMessageReturnInterface => {
+/**
+ * message 和 notification 公共方法
+ * 
+ * @param component 组件
+ * @returns 组件实例
+ */
+export const useMessage = (component: Component): UseMessageReturnInterface => {
   let seed = 1
 
-  /** 位置信息 */
-  const defaultOptions = {
-    message: { placement: 'top' },
-    notification: { placement: 'top-right' }
-  } as const
-
-  /** 组件实例 */
-  const componentVue = {
-    message: messageVue,
-    notification: notificationVue
-  }
-
+  /**
+   * 创建组件实例
+   * 
+   * @param options 传入的对象参数
+   * @returns 
+   */
   const instance: MessageFn & Partial<MessageFnWithType> = (options: MessageOptions): MessageInstance => {
+    /** 创建容器盒子 */
     const container: HTMLDivElement = document.createElement('div')
+    /** 每个 message 的唯一 id */
     const id = `message-${seed}`
 
+    /** 如果是字符串参数，那么直接将 message 赋值即可 */
     if (isString(options)) {
-      options = {
-        message: options
-      } as MessageOptions
+      options = { message: options }
     }
 
     const props: MessageOptions = {
       id,
-      ...defaultOptions[target],
-      ...options
+      ...{ placement: component.name === 'FMessage' ? 'top' : 'top-right' },
+      ...options,
+      /** 关闭动画结束时，移除 dom */
+      onDestroy: (): void => {
+        useRun(props.onClose)
+        render(null, container)
+      }
     } as const
 
-    /** 关闭动画结束时，移除 dom */
-    props.onDestroy = (): void => {
-      useRun(props.onClose)
-      render(null, container)
-    }
-
-    const VNode: VNode = createVNode(componentVue[target], props)
+    /**
+     * 创建一个 vNode
+     * 
+     * @see createVNode 
+     */
+    const VNode: VNode = createVNode(component, props)
 
     render(VNode, container)
+
     document.body.appendChild(container.firstElementChild as HTMLElement)
+
     const vm = VNode.component as ComponentInternalInstance
 
     seed++
+
     const instance: MessageInstance = massageManage.createInstance(
       {
         id,