+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- 修复 `FNotification` 组件名错误的问题
+- 优化 `FMessage` 和 `FNotification` 组件公共方法
+
 ## 0.20.0 (2023-01-17)
 
 - 优化 `f-space` 样式细节